Rework the main document.
Best practices suggest keeping each sentence/paragraph on a single line
to make it easier to track meaningful changes (turn on word wrap).

I added the missing image files.

I've decided to separate the history into a separate file; at some
point, we're going to need to generate a "diffs" document to show to the
board and there's no value in show history diffs (it might be a good
idea to move the revision information into the history.adoc file as
well). I'm thinking that we should populate the history section with the
entire document history (we can pull that from the earlier revisions).